Big changes - Put everything related to shell configuration in 'shell'
Other changes include:
- Made install_shell for installing everything related to bash and zsh
- Added functions and aliases that show current shell and help in
selecting a new default one
- Added a check on whether the packages bash-completions and
zsh-completion are installed
- Working towards removing check_completions, check_aliases and
check_keybinds
- Removed 00-rlwrapscripts
- Working towards removing any changes done to /root
- Removed duplicate files ranger.sh and flatpack.sh

+++ cli-tools/install_ast-grep.sh
@@ -1,3 +1,5 @@
+# https://github.com/ast-grep/ast-grep
+
 hash ast-grep &> /dev/null && SYSTEM_UPDATED='TRUE'
 
 TOP=$(git rev-parse --show-toplevel)
@@ -12,8 +14,6 @@ else
     . $TOP/checks/check_all.sh
 fi
 
-SCRIPT_DIR=$(get-script-dir)
-
 if ! hash ast-grep &>/dev/null; then
     if ! hash cargo &>/dev/null || ! [[ $PATH =~ '/.cargo/bin' ]]; then
         if ! test -f $TOP/cli-tools/pkgmngrs/install_cargo.sh; then
@@ -25,11 +25,11 @@ if ! hash ast-grep &>/dev/null; then
 
     cargo install ast-grep
 
-    if ! test -f ~/.bash_completion.d/ast-grep.bash; then
+    if test -d ~/.bash_completion.d/ && ! test -f ~/.bash_completion.d/ast-grep.bash; then
         echo "$(ast-grep completions bash)" >~/.bash_completion.d/ast-grep.bash
     fi
 
-    if ! test -f ~/.zsh_completion.d/ast-grep.zsh; then
+    if test -d ~/.zsh_completion.d/ && ! test -f ~/.zsh_completion.d/ast-grep.zsh; then
         echo "$(ast-grep completions zsh)" >~/.zsh_completion.d/ast-grep.zsh
     fi
 fi
